Simplify FRP path to set debug data.
